requirements.txt
scipy
scikit-learn
numpy
opencv_python
Pillow
torch>=1.2.0
torchvision
pyyaml
tqdm
tensorboardX
ftfy
regex
einops
timm
thop